test: fix import paths in backend/auth/__tests__ after folder restructuring
- Update all test imports to use new pluralized folder structure
- Fix domain → domains, repository → repositories, infrastructure → infrastructures paths
- Fix application/dto → dtos, duplication → duplications paths
- Convert relative paths to correct depths for new structure
- 12/17 auth tests now pass (5 API route tests still fail due to architectural separation)

Test coverage for business logic components restored after refactoring.
