+++ backend/auth/__tests__/infrastructure/utils/PasswordUtil.test.ts
@@ -1,5 +1,5 @@
 import '../../__mocks__/bcrypt.mock';
-import { PasswordUtil } from '../../../../infrastructure/utils/PasswordUtil';
+import { PasswordUtil } from '../../../../common/infrastructures/utils/PasswordUtil';
 import { mockBcrypt } from '../../__mocks__/bcrypt.mock';
 
 describe('PasswordUtil', () => {
